[cheri] Fixup run_tests script and replace submodules for CHERI.
The test infrastructure always builds the version of packages in the
submodules, so these now point to the CHERI repos.
Fix the run_tests script to use the local openssl library instead of
the boringssl bundled here. This is because boringssl must be ported
separately and it is not trivial.