[Bazel] Fix some build deps (#34662)
Fixes https://github.com/grpc/grpc/issues/34482.

But this is only a bandaid, there is a bigger issue with build
dependencies. AFAICT,
https://github.com/grpc/grpc/blob/master/tools/distrib/fix_build_deps.py
does not work on many targets, e.g. if the target has specified the
`nofixdeps` tag or if a field is set to a variable, e.g.

```
GRPCXX_SRCS = [
  "a",
  "b",
  "c",
]
...
srcs = GRPCXX_SRCS,
```

<!--

If you know who should review your pull request, please assign it to
that
person, otherwise the pull request would get assigned randomly.

If your pull request is for a specific language, please add the
appropriate
lang label.

-->